# backend/farms_openapi.json
{"openapi":"3.1.0","info":{"title":"Dr. Plant - Farm Service","version":"0.1.0"},"paths":{"/api/v1/farms":{"post":{"summary":"Create Farm","operationId":"create_farm_api_v1_farms_post","parameters":[{"name":"authorization","in":"header","required":false,"schema":{"anyOf":[{"type":"string"},{"type":"null"}],"title":"Authorization"}}],"requestBody":{"required":true,"content":{"application/json":{"schema":{"$ref":"#/components/schemas/FarmCreate"}}}},"responses":{"200":{"description":"Successful Response","content":{"application/json":{"schema":{"$ref":"#/components/schemas/FarmRead"}}}},"422":{"description":"Validation Error","content":{"application/json":{"schema":{"$ref":"#/components/schemas/HTTPValidationError"}}}}}},"get":{"summary":"List Farms","operationId":"list_farms_api_v1_farms_get","parameters":[{"name":"authorization","in":"header","required":false,"schema":{"anyOf":[{"type":"string"},{"type":"null"}],"title":"Authorization"}}],"responses":{"200":{"description":"Successful Response","content":{"application/json":{"schema":{"type":"array","items":{"$ref":"#/components/schemas/FarmRead"},"title":"Response List Farms Api V1 Farms Get"}}}},"422":{"description":"Validation Error","content":{"application/json":{"schema":{"$ref":"#/components/schemas/HTTPValidationError"}}}}}}},"/api/v1/fields":{"post":{"summary":"Create Field","operationId":"create_field_api_v1_fields_post","parameters":[{"name":"farm_id","in":"query","required":true,"schema":{"type":"string","format":"uuid","title":"Farm Id"}},{"name":"authorization","in":"header","required":false,"schema":{"anyOf":[{"type":"string"},{"type":"null"}],"title":"Authorization"}}],"requestBody":{"required":true,"content":{"application/json":{"schema":{"$ref":"#/components/schemas/FieldCreate"}}}},"responses":{"200":{"description":"Successful Response","content":{"application/json":{"schema":{"$ref":"#/components/schemas/FieldRead"}}}},"422":{"description":"Validation Error","content":{"application/json":{"schema":{"$ref":"#/components/schemas/HTTPValidationError"}}}}}}},"/api/v1/fields/{field_id}/crop-cycle":{"post":{"summary":"Create Crop Cycle","operationId":"create_crop_cycle_api_v1_fields__field_id__crop_cycle_post","parameters":[{"name":"field_id","in":"path","required":true,"schema":{"type":"string","format":"uuid","title":"Field Id"}},{"name":"authorization","in":"header","required":false,"schema":{"anyOf":[{"type":"string"},{"type":"null"}],"title":"Authorization"}}],"requestBody":{"required":true,"content":{"application/json":{"schema":{"$ref":"#/components/schemas/CropCycleCreateDTO"}}}},"responses":{"200":{"description":"Successful Response","content":{"application/json":{"schema":{"$ref":"#/components/schemas/CropCycleRead"}}}},"422":{"description":"Validation Error","content":{"application/json":{"schema":{"$ref":"#/components/schemas/HTTPValidationError"}}}}}}},"/api/v1/farms/{farm_id}/spraying-window":{"get":{"summary":"Get Spraying Window","operationId":"get_spraying_window_api_v1_farms__farm_id__spraying_window_get","parameters":[{"name":"farm_id","in":"path","required":true,"schema":{"type":"string","format":"uuid","title":"Farm Id"}},{"name":"authorization","in":"header","required":false,"schema":{"anyOf":[{"type":"string"},{"type":"null"}],"title":"Authorization"}}],"responses":{"200":{"description":"Successful Response","content":{"application/json":{"schema":{}}}},"422":{"description":"Validation Error","content":{"application/json":{"schema":{"$ref":"#/components/schemas/HTTPValidationError"}}}}}}},"/api/v1/fields/{field_id}/summary":{"get":{"summary":"Get Field Summary","operationId":"get_field_summary_api_v1_fields__field_id__summary_get","parameters":[{"name":"field_id","in":"path","required":true,"schema":{"type":"string","format":"uuid","title":"Field Id"}},{"name":"authorization","in":"header","required":false,"schema":{"anyOf":[{"type":"string"},{"type":"null"}],"title":"Authorization"}}],"responses":{"200":{"description":"Successful Response","content":{"application/json":{"schema":{"$ref":"#/components/schemas/FieldDetailedDTO"}}}},"422":{"description":"Validation Error","content":{"application/json":{"schema":{"$ref":"#/components/schemas/HTTPValidationError"}}}}}}},"/api/v1/fields/{field_id}/weather":{"get":{"summary":"Get Field Weather","operationId":"get_field_weather_api_v1_fields__field_id__weather_get","parameters":[{"name":"field_id","in":"path","required":true,"schema":{"type":"string","format":"uuid","title":"Field Id"}},{"name":"authorization","in":"header","required":false,"schema":{"anyOf":[{"type":"string"},{"type":"null"}],"title":"Authorization"}}],"responses":{"200":{"description":"Successful Response","content":{"application/json":{"schema":{}}}},"422":{"description":"Validation Error","content":{"application/json":{"schema":{"$ref":"#/components/schemas/HTTPValidationError"}}}}}}},"/api/v1/fields/{field_id}/soil-history":{"get":{"summary":"Get Soil History","operationId":"get_soil_history_api_v1_fields__field_id__soil_history_get","parameters":[{"name":"field_id","in":"path","required":true,"schema":{"type":"string","format":"uuid","title":"Field Id"}},{"name":"authorization","in":"header","required":false,"schema":{"anyOf":[{"type":"string"},{"type":"null"}],"title":"Authorization"}}],"responses":{"200":{"description":"Successful Response","content":{"application/json":{"schema":{}}}},"422":{"description":"Validation Error","content":{"application/json":{"schema":{"$ref":"#/components/schemas/HTTPValidationError"}}}}}}},"/api/v1/fields/{field_id}/satellite-gallery":{"get":{"summary":"Get Satellite Gallery","operationId":"get_satellite_gallery_api_v1_fields__field_id__satellite_gallery_get","parameters":[{"name":"field_id","in":"path","required":true,"schema":{"type":"string","format":"uuid","title":"Field Id"}},{"name":"authorization","in":"header","required":false,"schema":{"anyOf":[{"type":"string"},{"type":"null"}],"title":"Authorization"}}],"responses":{"200":{"description":"Successful Response","content":{"application/json":{"schema":{}}}},"422":{"description":"Validation Error","content":{"application/json":{"schema":{"$ref":"#/components/schemas/HTTPValidationError"}}}}}}},"/health":{"get":{"summary":"Health Check","operationId":"health_check_health_get","responses":{"200":{"description":"Successful Response","content":{"application/json":{"schema":{}}}}}}}},"components":{"schemas":{"CropCycleCreateDTO":{"properties":{"crop_name":{"type":"string","title":"Crop Name"},"variety":{"anyOf":[{"type":"string"},{"type":"null"}],"title":"Variety"},"planting_date":{"type":"string","format":"date-time","title":"Planting Date"},"expected_harvest_date":{"anyOf":[{"type":"string","format":"date-time"},{"type":"null"}],"title":"Expected Harvest Date"},"initial_quantity_seeds":{"anyOf":[{"type":"number"},{"type":"null"}],"title":"Initial Quantity Seeds"},"expected_duration_days":{"anyOf":[{"type":"integer"},{"type":"null"}],"title":"Expected Duration Days"}},"type":"object","required":["crop_name","planting_date"],"title":"CropCycleCreateDTO"},"CropCycleRead":{"properties":{"crop_name":{"type":"string","title":"Crop Name"},"variety":{"anyOf":[{"type":"string"},{"type":"null"}],"title":"Variety"},"planting_date":{"type":"string","format":"date-time","title":"Planting Date"},"expected_harvest_date":{"anyOf":[{"type":"string","format":"date-time"},{"type":"null"}],"title":"Expected Harvest Date"},"initial_quantity_seeds":{"anyOf":[{"type":"number"},{"type":"null"}],"title":"Initial Quantity Seeds"},"id":{"type":"string","format":"uuid","title":"Id"},"field_id":{"type":"string","format":"uuid","title":"Field Id"},"actual_harvest_date":{"anyOf":[{"type":"string","format":"date-time"},{"type":"null"}],"title":"Actual Harvest Date"},"status":{"$ref":"#/components/schemas/CropStatus"},"current_growth_stage":{"$ref":"#/components/schemas/GrowthStage"}},"type":"object","required":["crop_name","planting_date","id","field_id","status","current_growth_stage"],"title":"CropCycleRead"},"CropStatus":{"type":"string","enum":["PLANNED","ACTIVE","HARVESTED","FAILED"],"title":"CropStatus"},"EnvironmentalCacheRead":{"properties":{"id":{"type":"string","format":"uuid","title":"Id"},"field_id":{"type":"string","format":"uuid","title":"Field Id"},"ndvi_score":{"anyOf":[{"type":"number"},{"type":"null"}],"title":"Ndvi Score"},"soil_moisture_index":{"anyOf":[{"type":"number"},{"type":"null"}],"title":"Soil Moisture Index"},"accumulated_gdd":{"anyOf":[{"type":"number"},{"type":"null"}],"title":"Accumulated Gdd"},"evapotranspiration":{"anyOf":[{"type":"number"},{"type":"null"}],"title":"Evapotranspiration"},"captured_at":{"type":"string","format":"date-time","title":"Captured At"}},"type":"object","required":["id","field_id","captured_at"],"title":"EnvironmentalCacheRead"},"FarmCreate":{"properties":{"name":{"type":"string","title":"Name"},"village":{"anyOf":[{"type":"string"},{"type":"null"}],"title":"Village"},"district":{"anyOf":[{"type":"string"},{"type":"null"}],"title":"District"},"state":{"anyOf":[{"type":"string"},{"type":"null"}],"title":"State"}},"type":"object","required":["name"],"title":"FarmCreate"},"FarmRead":{"properties":{"name":{"type":"string","title":"Name"},"village":{"anyOf":[{"type":"string"},{"type":"null"}],"title":"Village"},"district":{"anyOf":[{"type":"string"},{"type":"null"}],"title":"District"},"state":{"anyOf":[{"type":"string"},{"type":"null"}],"title":"State"},"id":{"type":"string","format":"uuid","title":"Id"},"owner_id":{"type":"string","format":"uuid","title":"Owner Id"},"created_at":{"type":"string","format":"date-time","title":"Created At"},"fields":{"items":{"$ref":"#/components/schemas/FieldRead"},"type":"array","title":"Fields","default":[]}},"type":"object","required":["name","id","owner_id","created_at"],"title":"FarmRead"},"FieldCreate":{"properties":{"polygon":{"additionalProperties":true,"type":"object","title":"Polygon"},"soil_type_baseline":{"anyOf":[{"$ref":"#/components/schemas/SoilType"},{"type":"null"}]},"irrigation_type":{"anyOf":[{"$ref":"#/components/schemas/IrrigationType"},{"type":"null"}]}},"type":"object","required":["polygon"],"title":"FieldCreate"},"FieldDetailedDTO":{"properties":{"metadata":{"$ref":"#/components/schemas/FieldRead"},"active_crop":{"anyOf":[{"$ref":"#/components/schemas/CropCycleRead"},{"type":"null"}]},"latest_health":{"anyOf":[{"$ref":"#/components/schemas/EnvironmentalCacheRead"},{"type":"null"}]},"weather_forecast":{"items":{"additionalProperties":true,"type":"object"},"type":"array","title":"Weather Forecast","default":[]}},"type":"object","required":["metadata"],"title":"FieldDetailedDTO"},"FieldRead":{"properties":{"polygon":{"additionalProperties":true,"type":"object","title":"Polygon"},"soil_type_baseline":{"anyOf":[{"$ref":"#/components/schemas/SoilType"},{"type":"null"}]},"irrigation_type":{"anyOf":[{"$ref":"#/components/schemas/IrrigationType"},{"type":"null"}]},"id":{"type":"string","format":"uuid","title":"Id"},"farm_id":{"type":"string","format":"uuid","title":"Farm Id"},"agromonitoring_id":{"anyOf":[{"type":"string"},{"type":"null"}],"title":"Agromonitoring Id"},"area_hectares":{"anyOf":[{"type":"number"},{"type":"null"}],"title":"Area Hectares"}},"type":"object","required":["polygon","id","farm_id"],"title":"FieldRead"},"GrowthStage":{"type":"string","enum":["SEEDLING","VEGETATIVE","FLOWERING","RIPENING"],"title":"GrowthStage"},"HTTPValidationError":{"properties":{"detail":{"items":{"$ref":"#/components/schemas/ValidationError"},"type":"array","title":"Detail"}},"type":"object","title":"HTTPValidationError"},"IrrigationType":{"type":"string","enum":["DRIP","SPRINKLER","RAINFED","CANAL"],"title":"IrrigationType"},"SoilType":{"type":"string","enum":["CLAY","LOAMY","SANDY","BLACK","RED"],"title":"SoilType"},"ValidationError":{"properties":{"loc":{"items":{"anyOf":[{"type":"string"},{"type":"integer"}]},"type":"array","title":"Location"},"msg":{"type":"string","title":"Message"},"type":{"type":"string","title":"Error Type"},"input":{"title":"Input"},"ctx":{"type":"object","title":"Context"}},"type":"object","required":["loc","msg","type"],"title":"ValidationError"}}}}